#347 - Do not use PAT to request introspection (avoid exception related to missed `uma_protection` )
