--- oxd-server/src/main/java/org/gluu/oxd/server/model/UmaTokenFactory.java
@@ -1,20 +1,0 @@
-package org.gluu.oxd.server.model;
-
-import org.gluu.oxauth.model.uma.UmaScopeType;
-
-/**
- * @author Yuriy Zabrovarnyy
- */
-
-public class UmaTokenFactory {
-
-    private UmaTokenFactory() {
-    }
-
-    public static Token newToken(UmaScopeType scopeType) {
-        if (scopeType == UmaScopeType.PROTECTION) {
-            return new Pat();
-        }
-        throw new RuntimeException("Unknown scope type: " + scopeType);
-    }
-}
